Put generate function back into test file
