+++ hoomd/md/pytest/test_improper.py
@@ -5,13 +5,132 @@ import hoomd
 from hoomd.conftest import expected_loggable_params
 from hoomd.conftest import (logging_check, pickling_check,
                             autotuned_kernel_parameter_check)
-from hoomd.pytest.generate_periodic_improper_test_cases import (
-    get_force_vectors, periodic_improper_energy, chi_from_pos)
 
 import pytest
 import numpy
+import math
 
 import itertools
+
+
+# these series of functions are used to generate the test cases
+def dchi_dr1(n1, n2, r1, r2, r3, r4):
+    n1hat = n1 / numpy.sqrt(numpy.dot(n1, n1))
+    n2hat = n2 / numpy.sqrt(numpy.dot(n2, n2))
+    numerator = numpy.dot(n1hat, n2hat) * \
+        numpy.cross(numpy.dot(n1hat, n2hat) * n1hat - n2hat, r2 - r3) / \
+        numpy.linalg.norm(n1)
+
+    denominator = \
+        numpy.sqrt(
+            1 - numpy.dot(numpy.cross(n1hat, n2hat), numpy.cross(n1hat, n2hat))
+        ) \
+        * numpy.linalg.norm(numpy.cross(n1hat, n2hat))
+
+    return numerator / denominator
+
+
+def dchi_dr2(n1, n2, r1, r2, r3, r4):
+    n1hat = n1 / numpy.sqrt(numpy.dot(n1, n1))
+    n2hat = n2 / numpy.sqrt(numpy.dot(n2, n2))
+
+    numerator = numpy.dot(n1hat, n2hat) \
+        * (
+            numpy.cross(
+                numpy.dot(n1hat, n2hat) * n2hat - n1hat,
+                r3 - r4
+            )
+            / numpy.linalg.norm(n2) - numpy.cross(
+                numpy.dot(n1hat, n2hat) * n1hat - n2hat,
+                r1 - r3
+            ) / numpy.linalg.norm(n1)
+        )
+
+    denominator = numpy.sqrt(
+        1 - numpy.dot(numpy.cross(n1hat, n2hat), numpy.cross(n1hat, n2hat))) \
+        * numpy.linalg.norm(numpy.cross(n1hat, n2hat))
+
+    return numerator / denominator
+
+
+def dchi_dr3(n1, n2, r1, r2, r3, r4):
+    n1hat = n1 / numpy.sqrt(numpy.dot(n1, n1))
+    n2hat = n2 / numpy.sqrt(numpy.dot(n2, n2))
+
+    numerator = numpy.dot(n1hat, n2hat) \
+        * (
+            numpy.cross(
+                numpy.dot(n1hat, n2hat) * n1hat - n2hat,
+                r1 - r2
+            )
+            / numpy.linalg.norm(n1) - numpy.cross(
+                numpy.dot(n1hat, n2hat) * n2hat - n1hat,
+                r2 - r4
+            ) / numpy.linalg.norm(n2)
+        )
+
+    denominator = numpy.sqrt(
+        1 - numpy.dot(numpy.cross(n1hat, n2hat), numpy.cross(n1hat, n2hat))) \
+        * numpy.linalg.norm(numpy.cross(n1hat, n2hat))
+
+    return numerator / denominator
+
+
+def dchi_dr4(n1, n2, r1, r2, r3, r4):
+    n1hat = n1 / numpy.sqrt(numpy.dot(n1, n1))
+    n2hat = n2 / numpy.sqrt(numpy.dot(n2, n2))
+
+    numerator = numpy.dot(n1hat, n2hat) * \
+        numpy.cross(
+            numpy.dot(n1hat, n2hat) * n2hat - n1hat,
+            r2 - r3
+        ) / numpy.linalg.norm(n2)
+
+    denominator = \
+        numpy.sqrt(
+            1 - numpy.dot(
+                numpy.cross(n1hat, n2hat),
+                numpy.cross(n1hat, n2hat)
+            )
+        ) \
+        * numpy.linalg.norm(
+            numpy.cross(n1hat, n2hat)
+        )
+
+    return numerator / denominator
+
+
+def chi_from_pos(posa, posb, posc, posd):
+    n1 = numpy.cross(posa - posb, posb - posc)
+    n2 = numpy.cross(posb - posc, posc - posd)
+    mag = numpy.dot(n1, n2) / numpy.linalg.norm(n1) / numpy.linalg.norm(n2)
+    return math.acos(numpy.linalg.norm(mag))
+
+
+def du_dchi_periodic(chi, chi0, k, n, d):
+    return -k * n * d * numpy.sin(n * chi - chi0)
+
+
+def du_dchi_harmonic(chi, k, chi0):
+    return k * (chi - chi0)
+
+
+def periodic_improper_energy(chi, k, n, d, chi0):
+    return (k * (1 + d * numpy.cos(n * chi - chi0)))
+
+
+def get_force_vectors(chi, n1, n2, r1, r2, r3, r4, chi0, k, d, n):
+    f_matrix = numpy.zeros((4, 3))
+    f_matrix[0, :] = dchi_dr1(n1, n2, r1, r2, r3, r4) * du_dchi_periodic(
+        chi, chi0=chi0, k=k, d=d, n=n)
+    f_matrix[1, :] = dchi_dr2(n1, n2, r1, r2, r3, r4) * du_dchi_periodic(
+        chi, chi0=chi0, k=k, d=d, n=n)
+    f_matrix[2, :] = dchi_dr3(n1, n2, r1, r2, r3, r4) * du_dchi_periodic(
+        chi, chi0=chi0, k=k, d=d, n=n)
+    f_matrix[3, :] = dchi_dr4(n1, n2, r1, r2, r3, r4) * du_dchi_periodic(
+        chi, chi0=chi0, k=k, d=d, n=n)
+    return f_matrix
+
 
 # Test parameters include the class, improper params, force, and energy.
 # This is parameterized to plan for any future expansion with additional
@@ -47,7 +166,8 @@ improper_test_parameters = [
         0.007549784469704433,
     ),
     (hoomd.md.improper.Periodic, dict(k=3.0, d=-1, n=2, chi0=numpy.pi / 2),
-     get_force_vectors(n1=n1,
+     get_force_vectors(chi,
+                       n1=n1,
                        n2=n2,
                        r1=pos[0, :],
                        r2=pos[1, :],
@@ -59,7 +179,8 @@ improper_test_parameters = [
                        n=2) / 2,
      periodic_improper_energy(chi, k=3.0, d=-1, n=2, chi0=numpy.pi / 2) / 2),
     (hoomd.md.improper.Periodic, dict(k=10.0, d=1, n=1, chi0=numpy.pi / 4),
-     get_force_vectors(n1=n1,
+     get_force_vectors(chi,
+                       n1=n1,
                        n2=n2,
                        r1=pos[0, :],
                        r2=pos[1, :],
@@ -71,7 +192,8 @@ improper_test_parameters = [
                        n=1) / 2,
      periodic_improper_energy(chi, k=10.0, d=1, n=1, chi0=numpy.pi / 4) / 2),
     (hoomd.md.improper.Periodic, dict(k=5.0, d=1, n=3, chi0=numpy.pi / 6),
-     get_force_vectors(n1=n1,
+     get_force_vectors(chi,
+                       n1=n1,
                        n2=n2,
                        r1=pos[0, :],
                        r2=pos[1, :],
